Move filter pressed state inside Filter
