+++ src/utils/getTypesMock.js
@@ -1,0 +1,135 @@
+export const getTypesMock = () => [
+  {
+    id: 377,
+    scientific_name: 'Acacia',
+    name: 'Wattle',
+    parent_id: null,
+    count: 227,
+  },
+  {
+    id: 327,
+    scientific_name: 'Acacia aneura',
+    name: 'Mulga acacia',
+    parent_id: 377,
+    count: 1807,
+  },
+  {
+    id: 1236,
+    scientific_name: 'Acacia coriacea',
+    name: 'Wiry wattle',
+    parent_id: 377,
+    count: 5,
+  },
+  {
+    id: 979,
+    scientific_name: 'Acacia cultriformis',
+    name: 'Knife-leaf wattle',
+    parent_id: 377,
+    count: 4,
+  },
+  {
+    id: 175,
+    scientific_name: 'Acer',
+    name: 'Maple',
+    parent_id: null,
+    count: 2991,
+  },
+  {
+    id: 1347,
+    scientific_name: 'Acer circinatum',
+    name: 'Vine maple',
+    parent_id: 175,
+    count: 91,
+  },
+  {
+    id: 952,
+    scientific_name: 'Acer floridanum',
+    name: 'Florida maple',
+    parent_id: 175,
+    count: 9,
+  },
+  {
+    id: 944,
+    scientific_name: 'Acer negundo',
+    name: 'Box elder',
+    parent_id: null,
+    count: 13353,
+  },
+  {
+    id: 3482,
+    scientific_name: "Acer negundo 'Flamingo'",
+    name: 'Flamingo box elder',
+    parent_id: 944,
+    count: 2,
+  },
+  {
+    id: 3807,
+    scientific_name: "Acer negundo'Variegatum'",
+    name: 'Variegated box elder',
+    parent_id: 944,
+    count: 6,
+  },
+  {
+    id: 4401,
+    scientific_name: "Acer negundo'Wierii'",
+    name: 'Box elder',
+    parent_id: 944,
+    count: 1,
+  },
+  {
+    id: 1,
+    scientific_name: 'Acer negundo',
+    name: 'Box elder',
+    parent_id: null,
+    count: 13353,
+  },
+  {
+    id: 2,
+    scientific_name: "Acer negundo 'Flamingo'",
+    name: 'Flamingo box elder',
+    parent_id: 1,
+    count: 2,
+  },
+  {
+    id: 3,
+    scientific_name: "Acer negundo'Variegatum'",
+    name: 'Variegated box elder',
+    parent_id: 1,
+    count: 6,
+  },
+  {
+    id: 4,
+    scientific_name: "Acer negundo'Wierii'",
+    name: 'Box elder',
+    parent_id: 1,
+    count: 1,
+  },
+  {
+    id: 5,
+    scientific_name: "Acer negundo'Wierii'",
+    name: 'Box elder',
+    parent_id: 1,
+    count: 1,
+  },
+  {
+    id: 6,
+    scientific_name: "Acer negundo'Wierii'",
+    name: 'Box elder',
+    parent_id: 1,
+    count: 1,
+  },
+  {
+    id: 7,
+    scientific_name: "Acer negundo'Wierii'",
+    name: 'Box elder',
+    parent_id: 1,
+    count: 1,
+  },
+  {
+    id: 8,
+    scientific_name: "Acer negundo'Wierii'",
+    name: 'Box elder',
+    parent_id: 1,
+    count: 1,
+  },
+]
